requirements.txt
pandas==1.4.4
plotly==5.10.0
requests==2.28.1
bs4==0.0.1
matplotlib==3.5.2
plotly-express==0.4.1
yfinance==0.1.70
streamlit==1.9.2
tensorflow-cpu==2.9.2
keras==2.9.0
numpy==1.23.2
